requirements.txt
PyYAML==5.3.1
tbapy==1.3.2